added lint script and fixed linting errors
