+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # changelog
 
-## 1.0.1 release (10.03.23)
+## release (10.11.23)
+1. foregoing semver updates per template use
+2. created pr-prep yaml for github actions
+3. created issue and feature request forms
+4. added lint script and fixed linting errors
+
+## release (10.03.23)
 1. uncommented sentry code in `app.js`
 2. removed commented-out code across codebase
 3. removed unused files
 4. updated README
 5. added CHANGELOG
 
-## 1.0.0 release (10.02.23)
+## release (10.02.23)
 1. initialize bot and release boilerplate
