+++ package.json
@@ -8,6 +8,8 @@
   "license": "MIT",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "lint": "eslint . --ext .js",
+    "format": "eslint . --ext .js --fix",
     "start": "node --trace-warnings -r dotenv/config src/app/app.js"
   },
   "repository": {
