+++ src/app/utils/MongoDBUtils.js
@@ -3,30 +3,30 @@ import Log from './Log.js';
 
 const mongoDBURIPartial = `${process.env.MONGODB_PREFIX}://${process.env.MONGODB_USERNAME}:${process.env.MONGODB_PASS}@${process.env.MONGODB_CLUSTER}/`;
 const options = {
-    appname: process.env.APP_NAME,
-    serverAPI: {
-        version: ServerApiVersion.v1,
-        strict: true,
-        deprecationErrors: true
-    },
-    writeConcern: {
-        w: 'majority'
-    }
-}
+	appname: process.env.APP_NAME,
+	serverAPI: {
+		version: ServerApiVersion.v1,
+		strict: true,
+		deprecationErrors: true
+	},
+	writeConcern: {
+		w: 'majority'
+	}
+};
 
 const mongoClient = new MongoClient(mongoDBURIPartial, options);
 
 export default async function connect(database) {
-        try {
+	try {
 
-            Log.debug(`Connecting to ${database} for the first time!`);
+		Log.debug(`Connecting to ${database} for the first time!`);
 
-            await mongoClient.connect();
-            
-            await mongoClient.db(database).command({ ping: 1 });
-            Log.debug('Pinged your deployment. You successfully connected to MongoDB!');
+		await mongoClient.connect();
+			
+		await mongoClient.db(database).command({ ping: 1 });
+		Log.debug('Pinged your deployment. You successfully connected to MongoDB!');
 
-        } finally {
-            await mongoClient.close();
-        }
-    };
+	} finally {
+		await mongoClient.close();
+	}
+}
